# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "description": "An address that already exists is sent a link instead of being duplicated. If that account was never verified, these onboarding values overwrite the stashed ones.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted. The body is always empty and the response is byte-for-byte identical whether or not an account exists for the submitted address - that is the enumeration defense (REG-6, LOG-6), not an unfinished response."
          },
          "400": {
            "description": "Validation failed. `message` is the array of field errors produced by the global ValidationPipe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Refused by the throttler guard, which runs before the request body is ever validated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create an account and send its login link.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login-link": {
      "post": {
        "description": "An unknown address creates nothing and is sent nothing; only the response is identical. Every login link therefore belongs to a real user.",
        "operationId": "AuthController_requestLoginLink",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestLoginLinkDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted. The body is always empty and the response is byte-for-byte identical whether or not an account exists for the submitted address - that is the enumeration defense (REG-6, LOG-6), not an unfinished response."
          },
          "400": {
            "description": "Validation failed. `message` is the array of field errors produced by the global ValidationPipe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Refused by the throttler guard, which runs before the request body is ever validated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Send a login link to an existing account.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/verify": {
      "post": {
        "description": "First verification of an account also provisions its database, writes the profile the registration form described, and seeds the picked starter categories; a returning user just gets a session. **409** is the one actionable rejection: the link was replaced by a newer one, so the most recent email is the one to open. Every other dead token - unknown, expired, already spent, or belonging to a deleted account - is a 401.",
        "operationId": "AuthController_verify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyLoginLinkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed. `message` is the array of field errors produced by the global ValidationPipe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated. The bearer credential is missing, invalid, expired or already spent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "The request conflicts with the current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Refused by the throttler guard, which runs before the request body is ever validated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Spend a login link and start a session.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/session": {
      "get": {
        "description": "The central directory holds an email and a database pointer, so that is all this answers. Names, currency and budget live in the profile, which is its own endpoint.",
        "operationId": "AuthController_getSession",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated. The bearer credential is missing, invalid, expired or already spent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The identity behind a session token.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "description": "Revokes **only the bearer it is called with**, so other devices stay signed in. There is nothing to call twice: the token is dead afterwards, so a repeat answers **401** from the guard rather than a second 204. Clients should treat the local sign-out as their own responsibility - the frontend clears its session cookie whatever this answers, because a user who asked to leave must not be kept signed in by an unreachable API.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Signed out. No body."
          },
          "401": {
            "description": "Not authenticated. The bearer credential is missing, invalid, expired or already spent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "End this session.",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/assistant/messages": {
      "post": {
        "description": "Creates a conversation when `sessionId` is omitted and continues one when it is present, answering **201** with the stored question, the reply and the session both belong to. **Nothing is persisted unless the reply arrives**, so a failed, timed-out or cancelled turn stores neither the question nor an answer. `truncation` is non-null only when the account has more transactions than fit one prompt, and says how many of how many were sent and how far back the assistant could see. **503** means the assistant is not configured on this deployment, **504** that the model call did not finish in time - retrying the identical question is the right next move there, which is why it is distinct from the 503.",
        "operationId": "AssistantController_send",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendMessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed. `message` is the array of field errors produced by the global ValidationPipe.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated. The bearer credential is missing, invalid, expired or already spent.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No such resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Refused by the throttler guard, which runs before the request body is ever validated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "The feature is not configured on this deployment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "504": {
            "description": "A downstream call did not finish in time. Safe to retry.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Ask the assistant a question about your spending.",
        "tags": [
          "assistant"
        ]
      }
    },
...
    "/api/demo/session": {
      "post": {
        "description": "Hands the caller their own pre-seeded account out of a small pool, restoring it to the showcase fixture first unless it is untouched and was seeded today. No credential is required and none is accepted. **503** is the one ordinary rejection: every pooled account is leased to somebody else, and a later attempt will succeed as leases elapse. **404** means this deployment has the demo disabled.",
        "operationId": "DemoController_session",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoSessionResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No such resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited. Refused by the throttler guard, which runs before the request body is ever validated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "The feature is not configured on this deployment.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Lease a demo account and start a session on it.",
        "tags": [
          "demo"
        ]
      }
    }
  },
  "info": {
    "title": "Spendifico API",
    "description": "The HTTP contract of the Spendifico backend.\n\nGenerated from the NestJS code at build time, committed as `backend/openapi.json`, and\nthe only source of the frontend response types in `frontend/src/types/api.d.ts`.\nDo not edit either by hand: run `npm run api:sync` from the repo root.\n\nEvery operation can additionally answer **500** with `ErrorResponseDto`, via the global\nexception filter. It is stated once here rather than per operation: it is the same\nnon-actionable fact everywhere, and repeating it would widen every generated response type.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "The raw session token returned by POST /api/auth/verify. Opaque: 256 random bits, base64url, looked up by hash server-side."
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "description": "String for most errors; an array for class-validator's field messages.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "statusCode": {
            "type": "number"
          },
          "error": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "statusCode",
          "error",
          "timestamp",
          "path"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "currency": {
            "type": "string",
            "enum": [
              "EUR",
              "USD",
              "GBP"
            ],
            "default": "EUR",
            "description": "ISO 4217 code. Case-insensitive on the way in, stored and returned uppercase. Restricted to two-decimal currencies, because the API assumes an exponent of 2 everywhere.",
            "example": "EUR"
          },
          "monthlyBudget": {
            "type": "number",
            "description": "Major units (e.g. 2000.50). Stored as integer cents. The cap is not a\nproduct judgment: it keeps the cents conversion far inside JS safe-integer\nrange while staying generous for zero-decimal currencies, whose budgets\ncarry many digits.",
            "maximum": 1000000000,
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "monthStartDay": {
            "type": "integer",
            "description": "The day of the month you are paid on. Capped at 28 so the day exists in\nevery month.\n\n**This is the pay day, and onboarding asks for it as of PET-72.** It becomes\nthe account's first `period_rules` row, anchored to the most recent occurrence\nof this day - so the account starts with a real pay schedule rather than a\nsetting, and changing it later moves only the periods from the change onward.",
            "minimum": 1,
            "maximum": 28
          },
          "categories": {
            "description": "Ids from `GET /api/templates/categories`. May be empty. Unknown ids are a 400.",
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "fullName": {
            "type": "string",
            "description": "What the sidebar and greeting show. One field, not a first and last name.\n\nPET-72 collapsed the two. The app never used them apart - the sidebar wants\ninitials and a short name, both derivable from one string - so the second was\ndata collected in order to be thrown away. The form's label is \"Display name\",\nand a nickname is a legitimate answer.",
            "maxLength": 100
          },
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "monthlyBudget",
          "categories",
          "fullName",
          "email"
        ]
      },
      "RequestLoginLinkDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyLoginLinkDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "The raw token, exactly as it arrived in the link. 43 characters in practice\n(256 bits, base64url); the bound is loose enough not to encode that here\nwhile still keeping a megabyte body from ever reaching a hash function.",
            "maxLength": 128
          }
        },
        "required": [
          "token"
        ]
      },
      "VerifyResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "The raw session token, to be sent back as `Authorization: Bearer <token>`.\n\nThe only place it ever appears in a response - the server keeps its hash.\nThe frontend puts it in an httpOnly, first-party cookie and forwards it\nserver-side; it must never reach client-side JavaScript."
          },
          "expiresAt": {
            "type": "string",
            "description": "ISO 8601. Fixed at issue: using the session does not extend it."
          }
        },
        "required": [
          "token",
          "expiresAt"
        ]
      },
      "SessionResponseDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "description": "ISO 8601, the same instant verification returned."
          }
        },
        "required": [
          "userId",
          "email",
          "expiresAt"
        ]
      },
...
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "example": "How much did I spend on groceries last month?",
            "description": "The question. Trimmed, so a whitespace-only message is a 400 rather than a turn. The cap exists so a pasted novel cannot be what blows the model context; the composer restates it client-side, because `maxLength` reaches no generated type and the resulting 400 would otherwise produce advice the user cannot act on."
          },
          "sessionId": {
            "type": "string",
            "format": "uuid",
            "description": "The session to continue. Omit it to start one; the response carries the id that was created. An id naming no live session of yours is a **404**, unambiguously - this body references exactly one resource by id."
          }
        },
        "required": [
          "message"
        ]
      },
      "AssistantMessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "type": "string"
          },
          "content": {
            "type": "string",
            "example": "You spent 312.40 EUR on Groceries in August 2026."
          },
          "createdAt": {
            "type": "string",
            "description": "ISO 8601."
          }
        },
        "required": [
          "id",
          "role",
          "content",
          "createdAt"
        ]
      },
      "AssistantTruncationDto": {
        "type": "object",
        "properties": {
          "included": {
            "type": "number",
            "example": 3000,
            "description": "How many transactions were sent."
          },
          "total": {
            "type": "number",
            "example": 4210,
            "description": "How many the account has."
          },
          "oldestIncludedDate": {
            "type": "string",
            "example": "2024-02-09",
            "description": "The oldest date the assistant could see. Anything earlier is missing from its answer."
          }
        },
        "required": [
          "included",
          "total",
          "oldestIncludedDate"
        ]
      },
      "SendMessageResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "example": "How much did I spend on groceries?",
            "description": "Derived from the first message of the session and never rewritten."
          },
          "message": {
            "$ref": "#/components/schemas/AssistantMessageDto"
          },
          "reply": {
            "$ref": "#/components/schemas/AssistantMessageDto"
          },
          "truncation": {
            "nullable": true,
            "description": "Null whenever the whole history fitted, which is every account this project has.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AssistantTruncationDto"
              }
            ]
          }
        },
        "required": [
          "sessionId",
          "title",
          "message",
          "reply",
          "truncation"
        ]
      },
...
      "DemoSessionResponseDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "The raw session token, to be sent back as `Authorization: Bearer <token>`.\n\nThe only place it ever appears in a response - the server keeps its hash.\nThe frontend puts it in an httpOnly, first-party cookie and forwards it\nserver-side; it must never reach client-side JavaScript."
          },
          "expiresAt": {
            "type": "string",
            "description": "ISO 8601. Fixed at issue: using the session does not extend it.\n\n**This is the session's life, not the lease's.** A session runs for\n`SESSION_TTL_D` days while the lease runs for `DEMO_LEASE_TTL_M` minutes,\nso a visitor who leaves a tab open keeps a working session onto an account\nthat has since been handed to somebody else and rewritten under them. That\nis accepted rather than overlooked: shortening the session to the lease\nwould sign a visitor out mid-demo, and the data they would see afterwards is\nthe same fixture they started with."
          }
        },
        "required": [
          "token",
          "expiresAt"
        ]
      }
    }
  }
}